Update LICENSE file with commercial license information and GNU Affero General Public License version 3 (AGPLv3) details.
The LICENSE file has been updated to include information about obtaining a commercial license for commercial use of the software. The addition of the commercial license allows source code modifications and permits proprietary use without the copyleft provisions of the AGPLv3.

The commercial license is aimed at for-profit companies and individuals intending to use the software for commercial purposes, providing an alternative to using the software under the AGPLv3. The AGPLv3 allows users to use, modify, and distribute the software while preserving copyleft provisions for non-commercial projects.

This change clarifies the licensing terms for different use cases and ensures compliance with the appropriate licenses for commercial and non-commercial use.